# pkg/trace/agent: parallelize processor