@gbbr 2019-04-25
Applied changes in https://github.com/DataDog/datadog-agent/pull/3374